Fix some warnings for compiler

+++ trunk/src/utest/srs_utest_avc.cpp
@@ -247,7 +247,8 @@ VOID TEST(SrsAVCTest, H264IPBFrame)
         SrsRawH264Stream h; string frame;
         HELPER_ASSERT_SUCCESS(h.mux_ipb_frame((char*)"Hello", 5, frame));
         EXPECT_EQ(4+5, frame.length());
-        EXPECT_EQ(0, frame.at(0)); EXPECT_EQ(0, frame.at(1)); EXPECT_EQ(0, frame.at(2)); EXPECT_EQ(5, frame.at(3));
+        EXPECT_EQ(0, (uint8_t)frame.at(0)); EXPECT_EQ(0, (uint8_t)frame.at(1));
+        EXPECT_EQ(0, (uint8_t)frame.at(2)); EXPECT_EQ(5, (uint8_t)frame.at(3));
         EXPECT_STREQ("Hello", frame.substr(4).c_str());
     }
 }
